### 2025-09-19 #9028 @editorialbot
[PRE REVIEW]: r.earthworks: a GRASS tool for terrain modeling
**Submitting author:** @baharmon (<a href="http://orcid.org/0000-0002-6218-9318">Brendan Alexander Harmon</a>)
**Repository:** https://github.com/baharmon/r.earthworks
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.0
