### 2025-09-20 #9028 @baharmon
@kthyng

**Tests** There are automated tests in the [testsuite](https://github.com/baharmon/r.earthworks/tree/main/testsuite) directory. These can be run locally and are automatically run as part of the GRASS addons continuous integration.

**DOI**
The entries missing DOIs are a book and a software project that do not have identifiers.
