### 2025-10-13 489317b
updated readme, paper, and notebooks
