### 2025-10-16 #9028 @mengqi-z
👋 @micha-silver @angelmons @DejaBraz - Would any of you be willing to review this submission for the Journal of Open Source Software (JOSS)? JOSS uses a checklist-driven review process, which you can find here: https://joss.readthedocs.io/en/latest/review_criteria.html
