### 2025-10-17 #9270 @editorialbot
[REVIEW]: r.earthworks: a GRASS tool for terrain modeling
**Submitting author:** @baharmon (<a href="http://orcid.org/0000-0002-6218-9318">Brendan Alexander Harmon</a>)
**Repository:** https://github.com/baharmon/r.earthworks
**Branch with paper.md** (empty if default branch):
**Version:** 3.0.0
