### 2025-10-27 3d55747
pre-commit formatting
