+++ r.earthworks.py
@@ -287,11 +287,7 @@ def subdivision(coordinates, threshold, regions, cloud):
     # Discard empty quadrant
     else:
         gs.run_command(
-            "g.remove",
-            type="region",
-            name=region,
-            flags="f",
-            superquiet=True
+            "g.remove", type="region", name=region, flags="f", superquiet=True
         )
 
     # Check point count against threshold
@@ -308,10 +304,7 @@ def quadrant_nw(region, n, s, e, w, x, y):
 
     # Set region to north west quadrant
     gs.run_command(
-        "g.region",
-        n=n, s=n - y / 2, e=w + x / 2, w=w,
-        save=region,
-        overwrite=True
+        "g.region", n=n, s=n - y / 2, e=w + x / 2, w=w, save=region, overwrite=True
     )
 
 
@@ -320,10 +313,7 @@ def quadrant_ne(region, n, s, e, w, x, y):
 
     # Set region to north east quadrant
     gs.run_command(
-        "g.region",
-        n=n, s=n - y / 2, e=e, w=e - x / 2,
-        save=region,
-        overwrite=True
+        "g.region", n=n, s=n - y / 2, e=e, w=e - x / 2, save=region, overwrite=True
     )
 
 
@@ -332,10 +322,7 @@ def quadrant_sw(region, n, s, e, w, x, y):
 
     # Set region to south west quadrant
     gs.run_command(
-        "g.region",
-        n=s + y / 2, s=s, e=w + x / 2, w=w,
-        save=region,
-        overwrite=True
+        "g.region", n=s + y / 2, s=s, e=w + x / 2, w=w, save=region, overwrite=True
     )
 
 
@@ -344,10 +331,7 @@ def quadrant_se(region, n, s, e, w, x, y):
 
     # Set region to south east quadrant
     gs.run_command(
-        "g.region",
-        n=s + y / 2, s=s, e=e, w=e - x / 2,
-        save=region,
-        overwrite=True
+        "g.region", n=s + y / 2, s=s, e=e, w=e - x / 2, save=region, overwrite=True
     )
 
 
@@ -415,11 +399,7 @@ def prune(threshold, regions, cloud):
         # Discard regions
         else:
             gs.run_command(
-                "g.remove",
-                type="region",
-                name=region,
-                flags="f",
-                superquiet=True
+                "g.remove", type="region", name=region, flags="f", superquiet=True
             )
 
     return quadrants, coordinates
@@ -634,11 +614,7 @@ def grow_region(border, region, elevation):
 
     # Set expanded region
     gs.run_command(
-        "g.region",
-        n=n, s=s, e=e, w=w,
-        align=elevation,
-        save=region,
-        overwrite=True
+        "g.region", n=n, s=s, e=e, w=w, align=elevation, save=region, overwrite=True
     )
     gs.run_command("g.region", region=region)
 
