+++ README.md
@@ -44,18 +44,32 @@ r.contour input=earthworks output=contours step=2
 
 ## Tutorials
 
-**Basic earthworks** 
+Learn how model terrain with r.earthworks. 
+With this tool you can generate new terrain
+or modify existing terrain using cut and fill operations
+inspired by earthworking.
+The first tutorial on the basics of terrain modeling
+introduces key concepts such as cut and fill operations.
+The subsequent tutorials introduce more advanced concepts
+such as relative operations and growth and decay functions.
+These tutorials demonstrate how to generate terrain,
+model natural landforms,
+and model anthropogenic landforms - i.e. earthworks -
+such as roads, dams, and levees.
+Read our [paper](paper/paper.pdf) to learn more.
+
+[**Basic earthworks:**](https://grass-tutorials.osgeo.org/content/tutorials/earthworks/basics.html)
 Learn the basics of terrain modeling with r.earthworks.
 ![Basic earthworks](tutorials/images/basics_05.webp)
 
-**Gully modeling**
+[**Gully modeling:**](https://grass-tutorials.osgeo.org/content/tutorials/earthworks/gullies.html)
 Model gullies using relative cut operations.
 ![Gully modeling](tutorials/images/gullies_03.webp)
 
-**Coastal infrastructure**
+[**Coastal infrastructure:**](https://grass-tutorials.osgeo.org/content/tutorials/earthworks/levees.html)
 Model coastal infrastructure projects including levees and restored alluvial ridges.
 ![Coastal infrastructure](tutorials/images/levees_08.webp)
 
-**Terrain synthesis**
+[**Terrain synthesis:**](https://grass-tutorials.osgeo.org/content/tutorials/earthworks/synthesis.html)
 Synthesize terrain by applying landforms from one landscape to another.
 ![Terrain synthesis](tutorials/images/synthesis_07.webp)
