### 2025-11-10 #9270 @micha-silver
Hi @mengqi-z
I've gone over the paper, and it passes all the checklist requirements. However I was **not able to install** the addon. I posted an issue on the addon repo site for @baharmon (Brandon), so I'll hold off on giving a final review until I can get this worked out.
