+++ r.earthworks.py
@@ -217,6 +217,16 @@
 # % guisection: Quadtree
 # %end
 
+# %option G_OPT_M_NPROCS
+# % key: nprocs
+# % label: Number of threads for parallel computing
+# % description: 0: use OpenMP default; >0: use nprocs; <0: use MAX-nprocs
+# % required: no
+# % multiple: no
+# % answer: 0
+# % type: integer
+# %end
+
 # %flag
 # % key: p
 # % description: Print volume
@@ -755,6 +765,7 @@ def earthworking(
     cut,
     fill,
     disable,
+    threads
 ):
     """
     Model local earthworks
@@ -877,12 +888,15 @@ def earthworking(
                 f"{','.join(cut_operations)}"
                 f")"
                 f")",
+                nprocs=threads,
                 overwrite=True,
             )
         except:
             gs.fatal(
                 "Exceeded open file limit. "
-                "Try setting higher resource limits "
+                "Try enabling quadtree segmentation, "
+                "setting fewer threads for parallel computing, "
+                "or setting higher resource limits "
                 "with ulimit in the command line."
             )
 
@@ -900,12 +914,15 @@ def earthworking(
                 f"{','.join(fill_operations)}"
                 f")"
                 f")",
+                nprocs=threads,
                 overwrite=True,
             )
         except:
             gs.fatal(
                 "Exceeded open file limit. "
-                "Try setting higher resource limits "
+                "Try enabling quadtree segmentation, "
+                "setting fewer threads for parallel computing, "
+                "or setting higher resource limits "
                 "with ulimit in the command line."
             )
 
@@ -924,6 +941,7 @@ def earthworking(
                 f"{','.join(cut_operations)}"
                 f")"
                 f")",
+                nprocs=threads,
                 overwrite=True,
             )
 
@@ -939,17 +957,20 @@ def earthworking(
                 f"{','.join(fill_operations)}"
                 f")"
                 f")",
+                nprocs=threads,
                 overwrite=True,
             )
         except:
             gs.fatal(
                 "Exceeded open file limit. "
-                "Try setting higher resource limits "
+                "Try enabling quadtree segmentation, "
+                "setting fewer threads for parallel computing, "
+                "or setting higher resource limits "
                 "with ulimit in the command line."
             )
 
 
-def series(operation, function, cuts, fills, elevation, earthworks):
+def series(operation, function, cuts, fills, elevation, earthworks, threads):
     """
     Model cumulative earthworks
     """
@@ -972,7 +993,9 @@ def series(operation, function, cuts, fills, elevation, earthworks):
         )
         # Calculate net cut
         gs.mapcalc(
-            f"{earthworks} = if(isnull({cut}),{elevation},{cut})", overwrite=True
+            f"{earthworks} = if(isnull({cut}),{elevation},{cut})",
+            nprocs=threads,
+            overwrite=True
         )
 
     # Model net fill
@@ -990,7 +1013,9 @@ def series(operation, function, cuts, fills, elevation, earthworks):
         )
         # Calculate net fill
         gs.mapcalc(
-            f"{earthworks}= if(isnull({fill}),{elevation},{fill})", overwrite=True
+            f"{earthworks}= if(isnull({fill}),{elevation},{fill})",
+            nprocs=threads,
+            overwrite=True
         )
 
     # Model net cut and fill
@@ -1035,6 +1060,7 @@ def series(operation, function, cuts, fills, elevation, earthworks):
             # Calculate net elevation
             gs.mapcalc(
[…]
