+++ r.earthworks.md
@@ -195,6 +195,7 @@ operation set to `fill`,
 function set to `linear`,
 linear set to `0.25`,
 and flat set to `25`.
+Use flag `-p` to print the volume of fill.
 This will grade an embankment through the valley
 with a 50 meter wide roadway
 at a constant elevation of 95 meters
@@ -204,15 +205,13 @@ Optionally, compute contours with
 
 ```sh
 g.region n=217700 s=216200 w=639200 e=640700 res=10
-r.earthworks elevation=elevation earthworks=earthworks lines=roadsmajor z=95 function=linear linear=0.25 operation=fill flat=25
+r.earthworks elevation=elevation earthworks=earthworks lines=roadsmajor z=95 function=linear linear=0.25 operation=fill flat=25 -p
 r.contour input=earthworks output=contours step=2
 ```
 
 | Elevation | Earthworks |
 | --------- | ---------- |
 | ![Elevation](r_earthworks_07.png) | ![Earthworks](r_earthworks_08.png) |
-
-<!-- Print volume of fill -->
 
 When working with a large elevation raster,
 set the region to your area of interest
@@ -252,20 +251,35 @@ r.lake --overwrite elevation=earthworks water_level=104 lake=lake coordinates=63
 
 ## NOTES
 
-The current implementation of parallelization in r.mapcalc
-in the development version of GRASS
+In GRASS 8.5 and above,
+map algebra uses parallel computing
+for faster raster calculations. 
+The current implementation 
+of parallelization in r.mapcalc
 can open too many files
-causing r.earthworks to fail
+causing *r.earthworks* to fail
 for runs with multiple input coordinates
 depending on the open file limit of the user's system.
 This issue can be addressed
-by temporarily raising the open file limit:
+by enabling quadtree segmentation,
+setting fewer threads for parallel computing,
+or by raising the open file limit. 
+If this error occurs, try setting 
+smaller `threshold` and `border` parameters
+for quadtree segmentation.
+Alternatively, try setting 
+the number of threads for parallel computing
+with the `nprocs` parameter. 
+Use the shell command `ulimit`
+to temporarily raise the open file limit:
 
 ```bash
 ulimit -S -n 32768
 ```
 
 ## REFERENCES
+
+Harmon, B., Petrasova, A., and Petras, V. (2025). [r.earthworks: a GRASS tool for terrain modeling](https://github.com/baharmon/r.earthworks/blob/main/paper/paper.pdf) \[Preprint\].
 
 Harmon, B. (2025). r.earthworks (Version 2.0.0) \[Computer software\]. <https://doi.org/10.5281/zenodo.15507392>
 
