### 2026-01-28 #9270 @baharmon
@mengqi-z I have completed all the actions on the list.

Version: 2.0.0
DOI: [10.5281/zenodo.16846657](https://doi.org/10.5281/zenodo.16846657)

* [x]  Double check authors and affiliations (including ORCIDs).
* [x]  Make a release of the software with the latest changes from the review and post the version number here. This is the version that will be used in the JOSS paper: 2.0.
* [x]  Archive the released software on Zenodo or a similar service.
* [x]  Make sure that the archive (e.g., on Zenodo) has the correct metadata. This includes the title (which should match the paper title) and the author list (ensure the list is correct and does not include people who only made minor contributions). You may also add the authors' ORCIDs.
* [x]  Post the DOI of the archived version here: [10.5281/zenodo.16846657](https://doi.org/10.5281/zenodo.16846657)
