### 2026-01-28 #9270 @mengqi-z
@baharmon - I noticed that version 2.0.0 and its associated Zenodo DOI were published in August 2025, prior to the JOSS review. As part of our publication process, we ask authors to archive the final, reviewed version of the software. Could you please release the latest version of the software and post the corresponding version number and Zenodo DOI here? Thank you.
