### 2026-01-28 #9270 @mengqi-z
@baharmon - Yes. Any changes made since the start of the review should be included in the release/archive for JOSS publication, as reviewers evaluate the software with those updates when providing feedback.
