### 2026-01-28 #9270 @baharmon
@mengqi-z
Release: 3.0.0
DOI: [10.5281/zenodo.18407200](https://doi.org/10.5281/zenodo.18407200)
