### 2026-01-29 #9270 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
