### 2026-02-04 5363bd5
manuscript preparation for joss
