+++ paper/paper.bib
@@ -1,8 +1,8 @@
 @article{Hobley:2017,
 title={
-Creative computing with Landlab: an open-source toolkit for building,
+Creative computing with {Landlab}: an open-source toolkit for building,
 coupling, and exploring two-dimensional numerical models of
-Earth-surface dynamics
+{Earth}-surface dynamics
 },
 author={
 Hobley, Daniel EJ and Adams, Jordan M and Nudurupati, Sai Siddhartha and
@@ -19,7 +19,7 @@ doi={10.5194/esurf-5-21-2017}
 }
 
 @article{Barnhart:2020,
-title={Landlab v2. 0: a software package for Earth surface dynamics},
+title={Landlab v2. 0: a software package for {Earth} surface dynamics},
 author={
 Barnhart, Katherine R and Hutton, Eric WH and Tucker, Gregory E and
 Gasparini, Nicole M and Istanbulluoglu, Erkan and Hobley, Daniel EJ and
@@ -36,7 +36,15 @@ doi={10.5194/esurf-8-379-2020}
 }
 
 @software{GRASS,
-author = {{GRASS Development Team} and Landa, Martin and Neteler, Markus and Metz, Markus and Petrášová, Anna and Petráš, Vaclav and Clements, Glynn and Zigo, Tomáš and Larsson, Nicklas and Kladivová, Linda and Haedrich, Caitlin and Blumentrath, Stefan and Andreo, Veronica and Cho, Huidae and Gebbert, Sören and Nartišs, Māris and Kudrnovsky, Helmut and Delucchi, Luca and Zambelli, Pietro and Lennert, Moritz and Mitášová, Helena and Chemin, Yann and Pešek, Ondřej and Barton, Michael and Tawalika, Carmen and Ovsienko, Denis and Bowman, Hamish},
+author = {{
+GRASS Development Team} and Landa, Martin and Neteler, Markus and
+Metz, Markus and Petrášová, Anna and Petráš, Vaclav and Clements, Glynn and
+Zigo, Tomáš and Larsson, Nicklas and Kladivová, Linda and Haedrich, Caitlin and
+Blumentrath, Stefan and Andreo, Veronica and Cho, Huidae and Gebbert, Sören and
+Nartišs, Māris and Kudrnovsky, Helmut and Delucchi, Luca and Zambelli, Pietro and
+Lennert, Moritz and Mitášová, Helena and Chemin, Yann and Pešek, Ondřej and
+Barton, Michael and Tawalika, Carmen and Ovsienko, Denis and Bowman, Hamish
+},
 doi = {10.5281/zenodo.4621728},
 license = {GPL-2.0-or-later},
 title = {{GRASS GIS}},
@@ -46,8 +54,14 @@ version = {8.4.0}
 }
 
 @article{Haedrich:2023,
-author = {Haedrich, Caitlin and Petráš, Vaclav and Petrášová, Anna and Blumentrath, Stefan and Mitášová, Helena},
-title = {Integrating GRASS GIS and Jupyter Notebooks to facilitate advanced geospatial modeling education},
+author = {
+Haedrich, Caitlin and Petráš, Vaclav and Petrášová, Anna and 
+Blumentrath, Stefan and Mitášová, Helena
+},
+title = {
+Integrating {GRASS GIS} and {Jupyter Notebook}
+to facilitate advanced geospatial modeling education
+},
 journal = {Transactions in GIS},
 volume = {27},
 number = {3},
@@ -84,18 +98,25 @@ doi= {10.5194/gmd-12-2837-2019}
 }
 
 @article{Jasiewicz:2013,
+author = {Jarosław Jasiewicz and Tomasz F. Stepinski},
 title = {Geomorphons -- a pattern recognition approach to classification and mapping of landforms},
[…]
