+++ paper/paper.md
@@ -44,7 +44,7 @@ While modeling the shape of the earth’s surface is of interest to many sectors
 
 `r.earthworks` functionality includes the transformation of terrain rasters and calculation of volumetric change. Its features include cut and fill operations, relative or absolute datums, growth and decay functions for determining slopes, flats at local minima or maxima, and volumetric calculations. It can be used not only to model basic landforms such as peaks, ridges, slopes, valleys, and pits, but also complex natural and anthropogenic features. In `r.earthworks`, terrain – abstracted as a 2-dimensional manifold in 3-dimensional Euclidean space – is represented discretely as a raster grid for efficient storage, analysis, and transformation. Transformations are based on proposed local topographic extrema which can be derived from data, procedurally generated, or sketched. These extrema can be input as coordinates, points, lines, or a raster. The local minima and maxima are modeled as low points for cut operation or high points for fill operations. Transformations are a function of the existing elevation, change in vertical distance, and change in slope over horizontal distance. Vertical distance is calculated as the difference between proposed local extrema and a topographic datum, while change in slope is a function of growth and decay applied to horizontal change in distance. Growth and decay can be set to a linear, exponential, logistic, Gaussian, Cauchy-Lorentz, quadratic, or cubic function. The shape of the slope can be controlled by tuning the parameters of the chosen growth and decay function. Transformations are calculated independently for each local minima or maxima and are then accumulated before being applied to the existing terrain. 
 
-This tool was designed for use in spatial science workflows working with large terrain datasets. Since `r.earthworks` is raster based, it is efficient, scalable, flexible, and interoperable; it can process large elevation datasets, be used in workflows with other raster-based tools, and its results can be exported in common raster, point cloud, and array formats. After `r.earthworks` has been used to model topographic change, other `GRASS` tools can be used to analyze the resulting terrain and simulate physical processes such as surface flows of water and sediment. Through the `GRASS` Python application programming interface (API), `r.earthworks` can easily be integrated into data science workflows in Python. This tool includes automated tests, [documentation](https://grass.osgeo.org/grass-stable/manuals/addons/r.earthworks.html), and [tutorials](https://grass-tutorials.osgeo.org/content/tutorials/earthworks/earthworks.html) with accompanying computational notebooks. 
+This tool was designed for use in spatial science workflows working with large terrain datasets. Since `r.earthworks` is raster based, it is efficient, scalable, flexible, and interoperable; it can process large elevation datasets with quadtree partitioning, be used in workflows with other raster-based tools, and its results can be exported in common raster, point cloud, and array formats. After `r.earthworks` has been used to model topographic change, other `GRASS` tools can be used to analyze the resulting terrain and simulate physical processes such as surface flows of water and sediment. Through the `GRASS` Python application programming interface (API), `r.earthworks` can easily be integrated into data science workflows in Python. This tool includes automated tests, [documentation](https://grass.osgeo.org/grass-stable/manuals/addons/r.earthworks.html), and [tutorials](https://grass-tutorials.osgeo.org/content/tutorials/earthworks/earthworks.html) with accompanying computational notebooks. 
 
 # Usage
 
