### 2026-02-05 3195a4d
manuscript revision for joss
